Add base_resource_url to configs
In a proxied configuration Pedant needs to be aware of the different
between the place the server is listening at (platform.server) and the
resource that users actually contact. We frequently construct our
expected values based on the first, when the actual API will be
returning the second. We add a base_resource_url to the config and
platform data to enable us to distinquish that in tests.

Signed-off-by: Mark Anderson <mark@chef.io>

Fix pedant config (SQUASH)

Signed-off-by: Mark Anderson <mark@chef.io>